fix group edit mode actions
