+++ frontend/controls/peripherals/peripheral_form.tsx
@@ -2,7 +2,8 @@ import * as React from "react";
 import { PeripheralFormProps } from "./interfaces";
 import { sortResourcesById } from "../../util";
 import { Row, Col } from "../../ui";
-import { NameInputBox, PinDropdown, DeleteButton } from "../pin_form_fields";
+import { DeleteButton } from "../../ui/delete_button";
+import { NameInputBox, PinDropdown } from "../pin_form_fields";
 
 export const PeripheralForm = (props: PeripheralFormProps) =>
   <div className="peripheral-form">
