+++ frontend/farm_designer/map/background/selection_box_actions.tsx
@@ -3,7 +3,7 @@ import { TaggedPlant, AxisNumberProperty, Mode } from "../interfaces";
 import { SelectionBoxData } from "./selection_box";
 import { GardenMapState } from "../../interfaces";
 import { history } from "../../../history";
-import { selectPlant } from "../../actions";
+import { selectPlant } from "../actions";
 import { getMode } from "../util";
 
 /** Return all plants within the selection box. */
