+++ frontend/ui/index.ts
@@ -1,11 +1,22 @@
 export * from "./back_arrow";
 export * from "./blurable_input";
-export * from "./colors";
+export * from "./center_panel";
 export * from "./color_picker";
+export * from "./colors";
 export * from "./column";
+// export * from "./delete_button";
+export * from "./doc_link";
+export * from "./empty_state_wrapper";
+export * from "./expandable_header";
+export * from "./fallback_img";
+export * from "./fb_select";
 export * from "./help";
-export * from "./page";
+export * from "./input_error";
+export * from "./left_panel";
 export * from "./markdown";
+export * from "./new_fb_select";
+export * from "./page";
+export * from "./right_panel";
 export * from "./row";
 export * from "./saucer";
 export * from "./save_button";
@@ -14,10 +25,3 @@ export * from "./widget";
 export * from "./widget_header";
 export * from "./widget_footer";
 export * from "./widget_body";
-export * from "./fb_select";
-export * from "./new_fb_select";
-export * from "./fallback_img";
-export * from "./doc_link";
-export * from "./left_panel";
-export * from "./center_panel";
-export * from "./right_panel";
